* add slotted btree to repo

+++ tools/btree_benchmark.cpp
@@ -12,16 +12,15 @@ using namespace std;
 int main(int argc, char **argv)
 {
    uint32_t n = getenv("N") ? atoi(getenv("N")) : 10e6;
-   uint32_t threads =  getenv("T") ? atoi(getenv("T")) : 20;
+   uint32_t threads = getenv("T") ? atoi(getenv("T")) : 20;
    libgcc::memory_block = new u8[4 * 1024 * n]; // wasteful
-   std::memset(libgcc::memory_block,0,4 * 1024 * n );
+   std::memset(libgcc::memory_block, 0, 4 * 1024 * n);
    libgcc::memory_head = 0;
 
-   std::vector<uint32_t> work(n);
-   for ( uint32_t i = 0; i < n; i++ )
+   std::vector<u64> work(n);
+   for ( u64 i = 0; i < n; i++ )
       work[i] = i;
    std::random_shuffle(work.begin(), work.end());
-
 
 
    tbb::task_scheduler_init taskScheduler(threads);
@@ -88,7 +87,7 @@ int main(int argc, char **argv)
 //   }
    {
 
-      optimistic::BTree<uint32_t, uint32_t> btree;
+      optimistic::BTree<u64, u64> btree;
       // insert
       {
          PerfEventBlock b(e, n);
@@ -104,7 +103,7 @@ int main(int argc, char **argv)
          });
       }
 
-      if(getenv("ALL")){
+      if ( getenv("ALL")) {
          // lookup
          {
             PerfEventBlock b(e, n);
@@ -112,9 +111,9 @@ int main(int argc, char **argv)
             e.setParam("approach", "optimistic");
             e.setParam("threads", threads);
 
-            tbb::parallel_for(tbb::blocked_range<uint32_t>(0, n), [&](const tbb::blocked_range<uint32_t> &range) {
-               for ( uint32_t i = range.begin(); i < range.end(); i++ ) {
-                  uint32_t result;
+            tbb::parallel_for(tbb::blocked_range<u64>(0, n), [&](const tbb::blocked_range<u64> &range) {
+               for ( u64 i = range.begin(); i < range.end(); i++ ) {
+                  u64 result;
                   bool success = btree.lookup(work[i], result);
                   assert(success && result == work[i]);
                }
@@ -130,10 +129,10 @@ int main(int argc, char **argv)
             e.setParam("threads", threads);
 
             tbb::parallel_for(tbb::blocked_range<uint32_t>(0, n), [&](const tbb::blocked_range<uint32_t> &range) {
-               uint32_t sum = 0;
-               for ( uint32_t i = range.begin(); i < range.end(); i++ ) {
+               u64 sum = 0;
+               for ( u64 i = range.begin(); i < range.end(); i++ ) {
                   if ( i % 10 < 6 ) {
-                     uint32_t result;
+                     u64 result;
                      bool success = btree.lookup(work[i], result);
                      if ( success ) {
                         assert(result == work[i]);
